+++ users/src/sdk/models/shared/accountevents.ts
@@ -7,6 +7,10 @@ import { Expose } from "class-transformer";
 
 export class AccountEvents extends SpeakeasyBase {
     @SpeakeasyMetadata()
+    @Expose({ name: "failedVerifyMicroDepositAt" })
+    failedVerifyMicroDepositAt?: string[];
+
+    @SpeakeasyMetadata()
     @Expose({ name: "initialSetupAt" })
     initialSetupAt?: string;
 
@@ -15,6 +19,18 @@ export class AccountEvents extends SpeakeasyBase {
     isInitialSetupStarted?: boolean;
 
     @SpeakeasyMetadata()
+    @Expose({ name: "requestedMicroDepositAt" })
+    requestedMicroDepositAt?: string;
+
+    @SpeakeasyMetadata()
+    @Expose({ name: "sentMicroDepositEmailAt" })
+    sentMicroDepositEmailAt?: string[];
+
+    @SpeakeasyMetadata()
     @Expose({ name: "transactionsSyncedAt" })
     transactionsSyncedAt?: string;
+
+    @SpeakeasyMetadata()
+    @Expose({ name: "verifiedMicroDepositAt" })
+    verifiedMicroDepositAt?: string;
 }