+++ users/src/sdk/sdk.ts
@@ -23,11 +23,11 @@ export const ServerList = [
     /**
      * Wingspan's Users API in Production
      */
-    "https://api.wingspan.app/users",
+    "https://api.wingspan.app",
     /**
      * Wingspan's Users API in Staging
      */
-    "https://stagingapi.wingspan.app/payments",
+    "https://stagingapi.wingspan.app",
 ] as const;
 
 /**
@@ -63,9 +63,9 @@ export class SDKConfiguration {
     serverDefaults: any;
     language = "typescript";
     openapiDocVersion = "1.0.0";
-    sdkVersion = "2.0.2";
-    genVersion = "2.195.2";
-    userAgent = "speakeasy-sdk/typescript 2.0.2 2.195.2 1.0.0 @wingspan/users";
+    sdkVersion = "2.0.3";
+    genVersion = "2.209.0";
+    userAgent = "speakeasy-sdk/typescript 2.0.3 2.209.0 1.0.0 @wingspan/users";
     retryConfig?: utils.RetryConfig;
     public constructor(init?: Partial<SDKConfiguration>) {
         Object.assign(this, init);