+++ users/RELEASES.md
@@ -39,3 +39,13 @@ Based on:
 - [typescript v2.0.2] users
 ### Releases
 - [NPM v2.0.2] https://www.npmjs.com/package/@wingspan/users/v/2.0.2 - users
+
+## 2023-12-04 01:30:47
+### Changes
+Based on:
+- OpenAPI Doc 1.0.0 https://docs.wingspan.app/openapi/6470eea6a7577e003be20a26
+- Speakeasy CLI 1.125.0 (2.209.0) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [typescript v2.0.3] users
+### Releases
+- [NPM v2.0.3] https://www.npmjs.com/package/@wingspan/users/v/2.0.3 - users